requirements.txt
wordcloud
pyecharts
pylint
requests
yapf
jieba
matplotlib
numpy
pandas
pyecharts-snapshot 
pillow
